# Use mach_absolute_time as a timer on macOS