# Performance comparison
## Before (gettimeofday)
```
In [1]: import wsgi_lineprof.extensions

In [2]: g = wsgi_lineprof.extensions.LineProfiler().get_timer

In [3]: %timeit g()
110 ns ± 4.72 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [4]: %timeit g()
111 ns ± 2.2 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [5]: %timeit g()
119 ns ± 7.48 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

## After (get_absolute_time)
```
In [1]: import wsgi_lineprof.extensions

In [2]: g = wsgi_lineprof.extensions.LineProfiler().get_timer

In [3]: %timeit g()
107 ns ± 8.48 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [4]: %timeit g()
98.5 ns ± 4.26 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [5]: %timeit g()
107 ns ± 11.3 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```